fix: auto padding issue
